% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{standardize_division}
\alias{standardize_division}
\title{Standardize Police Division Names}
\usage{
standardize_division(data, division_col = "division")
}
\arguments{
\item{data}{A data frame or tibble containing a division column.}

\item{division_col}{The name of the division column as a string.}
}
\value{
A tibble with the specified division column standardized and
  converted to a factor.
}
\description{
Maps known variations of Dallas Police division names to a standard set and
converts the column to a factor.
}
